// types/inventoryVoucher/PurchaseVoucherTypes.d.ts
export interface PurchaseVoucherFormData {
  purchaseType: string;
  orderPurchaseType: string;
  purchaseDate: Date;
  partyId: string;
  purchaseNo: string;
  itemId: string;
  quantity: string;
  rate: string;
  roundOff: string;
  discount: string;
  transMode: string;
  bankId: string;
}
// ...
export interface PurchaseTableData {
  purchaseDate: Date;
  partyId: string;
  partyName: string;
  purchaseNo: string;
  itemId: string;
  itemName: string;
  quantity: string;
  rate: string;
  taxableTotal: number;
  cgst: string | number;
  sgst: string | number;
  grandTotal: number;
  itemGl: string | number;
  orderPurchaseType: string;
}

export interface ItemRequisitionRow {
  Id: number | string;
  Item_Id?: number | string;
  Item_Name?: string;
  Item_Qnty?: string | number;
  Qnty?: string | number;
  Quantity?: string | number;
  Item_Rate?: string | number;
  Rate?: string | number;
  Pur_Rate?: string | number;
  Order_No?: string;
  Req_No?: string;
  Party_Name?: string;
  Req_Id?: number | string;
  Purchase_Gl?: string | number;
  Row_Key?: string;
  ItemRow?: ItemRequisitionRow[];
}