// src/types/master/ItemTypes.d.ts
export interface ItemTableData {
  Id: number;
  Cat_Id: numberl;
  Model_Id: number | null;
  Size_Id: number | null;
  Color_Id: number | null;
  Item_Name: string;
  Item_Sh_Name: string;
  Unit_Id: number;
  Purchase_Gl: number;
  Sales_Gl: number;
  CGST: string;
  SGST: string;
  IGST: string;
  Pur_Rate: string;
  Sale_Rate: string;
  Cat_Name: string;
  Open_Qnty: string;
  Item_Rate: string;
}